% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columnsummary.R
\name{col_sum}
\alias{col_sum}
\title{Summary on data frame columns.}
\usage{
col_sum(x = mpg, f = mean)
}
\arguments{
\item{x}{The datasets the user is interested in.}

\item{f}{The summary function the user is interested in.}
}
\description{
This function allows to do summary on
all the numeric columns in a data frame.
}
\details{
The function asks the user to pick one data frame,
and the summary type he or she is interested in.
Then the user can applies a summary function to
every numeric column in a selected data frame and
get the analysis results quickly.
}
\examples{
col_sum(iris, median)
}
